src/third_party/blink/web_tests/external/wpt/service-workers/service-worker/update.https-expected.txt

This is a testharness.js-based test.
PASS update() should succeed when new script is available.
PASS update() should fail when mime type is invalid.
FAIL update() should fail when a response for the main script is redirect. assert_throws: function "function() { throw e }" threw object "SecurityError: Failed to update a ServiceWorker: The script resource is behind a redirect, which is disallowed." ("SecurityError") expected object "TypeError" ("TypeError")
PASS update() should fail when a new script contains a syntax error.
PASS update() should resolve when the install event throws.
PASS update() should fail when the pending uninstall flag is set.
PASS update() should succeed when the script shrinks.
Harness: the test ran to completion.
